Restructure worker code to be a bit more organized, and rename EventIncidents to ShareInstance
