+++ worker/sync/src/utils/crypto.ts
